api/requirements.txt
aiosqlite==0.19.0
anyio==3.6.2
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
certifi==2022.12.7
click==8.1.3
comm==0.1.3
databases==0.7.0
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
executing==1.2.0
fastapi==0.95.1
flake8==6.0.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
ipykernel==6.23.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
nest-asyncio==1.5.6
orjson==3.8.11
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.15.1
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
pyzmq==25.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
stack-data==0.6.2
starlette==0.26.1
tomli==2.0.1
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
